+++ app/src/main/java/io/legado/app/data/entities/Book.kt
@@ -39,7 +39,7 @@ data class Book(
     var customIntro: String? = null,            // 简介内容(用户修改)
     var charset: String? = null,                // 自定义字符集名称(仅适用于本地书籍)
     var type: Int = 0,                          // 0:text 1:audio
-    var group: Int = 0,                         // 自定义分组索引号
+    var group: Long = 0,                         // 自定义分组索引号
     var latestChapterTitle: String? = null,     // 最新章节标题
     var latestChapterTime: Long = System.currentTimeMillis(),            // 最新章节标题更新时间
     var lastCheckTime: Long = System.currentTimeMillis(),                // 最近一次更新书籍信息的时间
